refactor: reusable validator schema
