+++ src/utils/constants.jsx
@@ -1,0 +1,6 @@
+export const navigation = [
+    { name: "home", href: "/" },
+    { name: "blogs", href: "#" },
+    { name: "about", href: "#" },
+    { name: "contactus", href: "#" },
+];
